requirements.txt
Django==1.11.11
django-crispy-forms==1.6.1
django-formtools==2.0
httplib2==0.10.3
Pillow==4.0.0
PyMySQL==0.7.10
django-simple-captcha==0.4.6
django-pure-pagination
django-debug-toolbar
git+git://github.com/sshwsfc/xadmin.git@django2